requirements.txt
streamlit
diffusers
transformers
accelerate
safetensors
torch
python-dotenv
